chore: fixed issues with dropdown detecting as clicking outside modal
