+++ packages/p2p/src/components/modal-manager/modals/create-ad-add-payment-method-modal/index.js
@@ -1,0 +1,4 @@
+import CreateAdAddPaymentMethodModal from './create-ad-add-payment-method-modal.jsx';
+import './create-ad-add-payment-method-modal.scss';
+
+export default CreateAdAddPaymentMethodModal;
