+++ packages/p2p/src/components/my-ads/create-ad-form-payment-methods.jsx
@@ -5,11 +5,13 @@ import PaymentMethodCard from '../my-profile/payment-methods/payment-method-card
 import { localize } from 'Components/i18next';
 import BuyAdPaymentMethodsList from './buy-ad-payment-methods-list.jsx';
 import SellAdPaymentMethodsList from './sell-ad-payment-methods-list.jsx';
+import { useModalManagerContext } from 'Components/modal-manager/modal-manager-context';
 
 const CreateAdFormPaymentMethods = ({ is_sell_advert, onSelectPaymentMethods }) => {
     const { my_ads_store, my_profile_store } = useStores();
     const [selected_buy_methods, setSelectedBuyMethods] = React.useState([]);
     const [selected_sell_methods, setSelectedSellMethods] = React.useState([]);
+    const { showModal } = useModalManagerContext();
 
     const onClickPaymentMethodCard = payment_method => {
         if (!my_ads_store.payment_method_ids.includes(payment_method.ID)) {
@@ -49,7 +51,11 @@ const CreateAdFormPaymentMethods = ({ is_sell_advert, onSelectPaymentMethods }) 
             return (
                 <SellAdPaymentMethodsList
                     selected_methods={selected_sell_methods}
-                    onClickAdd={() => my_ads_store.setShouldShowAddPaymentMethodModal(true)}
+                    onClickAdd={() =>
+                        showModal({
+                            key: 'CreateAdAddPaymentMethodModal',
+                        })
+                    }
                     onClickPaymentMethodCard={onClickPaymentMethodCard}
                 />
             );
@@ -60,7 +66,11 @@ const CreateAdFormPaymentMethods = ({ is_sell_advert, onSelectPaymentMethods }) 
                 is_add
                 label={localize('Payment method')}
                 medium
-                onClickAdd={() => my_ads_store.setShouldShowAddPaymentMethodModal(true)}
+                onClickAdd={() =>
+                    showModal({
+                        key: 'CreateAdAddPaymentMethodModal',
+                    })
+                }
             />
         );
     }
