+++ packages/p2p/src/components/my-profile/payment-methods/add-payment-method/add-payment-method-form.jsx
@@ -10,8 +10,8 @@ import ModalForm from 'Components/modal-manager/modal-form';
 import { useModalManagerContext } from 'Components/modal-manager/modal-manager-context';
 
 const AddPaymentMethodForm = ({ should_show_separated_footer = false }) => {
-    const { my_ads_store, my_profile_store } = useStores();
-    const { showModal } = useModalManagerContext();
+    const { my_profile_store } = useStores();
+    const { hideModal, showModal } = useModalManagerContext();
 
     React.useEffect(() => {
         my_profile_store.getPaymentMethodsList();
@@ -104,7 +104,7 @@ const AddPaymentMethodForm = ({ should_show_separated_footer = false }) => {
                                             });
                                         } else {
                                             my_profile_store.hideAddPaymentMethodForm();
-                                            my_ads_store.setShouldShowAddPaymentMethodModal(false);
+                                            hideModal();
                                         }
                                     }}
                                     type='button'
