+++ packages/p2p/src/constants/modals.js
@@ -1,6 +1,16 @@
 import React from 'react';
 
 export const modals = {
+    AdExceedsDailyLimitModal: React.lazy(() =>
+        import(
+            /* webpackChunkName: "ad-exceeds-daily-limit-modal" */ 'Components/modal-manager/modals/ad-exceeds-daily-limit-modal'
+        )
+    ),
+    AddPaymentMethodErrorModal: React.lazy(() =>
+        import(
+            /* webpackChunkName: "add-payment-method-error-modal" */ 'Components/modal-manager/modals/add-payment-method-error-modal'
+        )
+    ),
     BuySellModal: React.lazy(() =>
         import(/* webpackChunkName: "buy-sell-modal" */ 'Components/modal-manager/modals/buy-sell-modal')
     ),
@@ -9,9 +19,9 @@ export const modals = {
             /* webpackChunkName: "cancel-add-payment-method-modal" */ 'Components/modal-manager/modals/cancel-add-payment-method-modal'
         )
     ),
-    AdExceedsDailyLimitModal: React.lazy(() =>
+    CreateAdAddPaymentMethodModal: React.lazy(() =>
         import(
-            /* webpackChunkName: "ad-exceeds-daily-limit-modal" */ 'Components/modal-manager/modals/ad-exceeds-daily-limit-modal'
+            /* webpackChunkName: "create-ad-add-payment-method-modal" */ 'Components/modal-manager/modals/create-ad-add-payment-method-modal'
         )
     ),
     EmailLinkVerifiedModal: React.lazy(() =>
@@ -22,13 +32,13 @@ export const modals = {
     FilterModal: React.lazy(() =>
         import(/* webpackChunkName: "filter-modal" */ 'Components/modal-manager/modals/filter-modal')
     ),
+    MyAdsDeleteModal: React.lazy(() =>
+        import(/* webpackChunkName: "my-ads-delete-modal" */ 'Components/modal-manager/modals/my-ads-delete-modal')
+    ),
     MyAdsDeleteErrorModal: React.lazy(() =>
         import(
             /* webpackChunkName: "my-ads-delete-error-modal" */ 'Components/modal-manager/modals/my-ads-delete-error-modal'
         )
-    ),
-    MyAdsDeleteModal: React.lazy(() =>
-        import(/* webpackChunkName: "my-ads-delete-modal" */ 'Components/modal-manager/modals/my-ads-delete-modal')
     ),
     MyAdsFloatingRateSwitchModal: React.lazy(() =>
         import(
